+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 8.1.5 (git master)
+* Ensure that the summary flag is true for WBS items, even if they have no child activities.
 
 ## 8.1.4 (31/08/2020)
 * Fix CVE-2020-25020: XXE vulnerability (with thanks to Sangeetha Rajesh S, ZOHO Corporation)